requirements.txt
Flask==2.3.2
Flask-RESTful==0.3.10
GitPython==3.1.32
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.6